Support yj's updated download filename
As of `yj` 5.1.0+ the filename of the asset attached to the GitHub release
is now `yj-linux-arm64` instead of `yj-linux`.

This change ensures that the new yj default (as of #164) is able to be
downloaded, since currently the action fails with:

```
Installing yj 5.1.0
curl: (22) The requested URL returned error: 404
Error: Process completed with exit code 22.
```

The bash string comparison of the semver version isn't ideal, but it
doesn't feel worth adding further complexity given that:
- all `yj` releases so far use a full semver version (and not say an `X.Y` version)
- `yj` major releases are infrequent, so it's going to be a while before v10 is out
  (which would break the comparison) - by which point this backwards compatibiluty
  comparison can just be removed (and support for pre 5.1.0 removed).

Signed-off-by: Ed Morley <501702+edmorley@users.noreply.github.com>