Increase timeouts for actions that rely on actions (again)
In #214 the timeouts for Actions that rely on other Actions was
bumped from 2 minutes to 8 minutes.

However, there are still occasionally instances where the registry
request issue isn't processed for up to 15-16 minutes due to delays
in the GitHub Actions runner being started.

As such this raises the timeout again, from 8 minutes to 20 minutes.

Signed-off-by: Ed Morley <501702+edmorley@users.noreply.github.com>